few fixes

+++ api/src/auth/auth.service.ts
@@ -12,7 +12,7 @@ export class AuthService {
   constructor(
     protected readonly usersService: UsersService,
     protected readonly jwtService: JwtService,
-  ) {}
+  ) { }
 
   public async validateUser(email: string, password: string) {
     const user = await this.usersService.findByEmail(email);
@@ -23,12 +23,15 @@ export class AuthService {
     }
   }
 
-  public async login(user: User) {
+  public async login(user: User, token?: RefreshToken) {
     const payload = { id: user.id };
-    const refreshToken = uuid();
+    const refreshToken = token ? token.token : uuid();
 
-    await RefreshToken.query().delete().where({ userId: user.id });
-    await RefreshToken.query().insert({ userId: user.id, token: refreshToken });
+    if (!token) {
+      const refreshToken = uuid();
+      await RefreshToken.query().delete().where({ userId: user.id });
+      await RefreshToken.query().insert({ userId: user.id, token: refreshToken });
+    }
 
     return {
       accessToken: this.jwtService.sign(payload),
@@ -48,8 +51,7 @@ export class AuthService {
     }
 
     const user = await this.usersService.findById(refreshToken.userId);
-
-    return this.login(user);
+    return this.login(user, refreshToken);
   }
 
   public async getUser(id: number) {
